Used for numerical analysis of new data
This is the Boston Housing data set, broken into two parts. Boston_Housing is rows 6:end of file. The NewBoston.csv file are the first five rows of the Boston Housing data set. The goal here is to make models and predictions on Boston_Housing, and then use those models to predict on NewBoston.csv.